Fix southpanel being collapsed when toggling multiple gfi and single gfi was already active
Two events are generated, a toggleX event with state=true, then
a toggle even with state=false. In certain conditions (depending on
the active layer, the order of layers..) the check was trigerring
and collapsing the southpanel. Make sure that it is only triggered
if there's only one layer in ctrl.layers (ie we're not going to
do a multiple gfi)