mapfishapp - upload / download geo file (#754)
First naïve implementation of the download feature, all the security is
based on the fact that only zip files and xml derivatives can be used as
geo files. This might deserve a better hardenning to avoid security
flaws, but should address the feature request for now.

This seems not really functional with geotools implementation, so it
might need a gdal/ogr-enabled mapfishapp.

Test: mainly manual, might need to improve / adapt test suite.