debian packaging - debian:deploy can fail
If the debian:build target generates a deb package at minute n and the
deb:deploy target tries to publish the package on the repository at
minute n+1, the deployment does not succeed silentlyi because the deb
file is calculated (in case of snapshot builds) given the current time.

Using a generated modification time of a file (as target/) shall solve
the issue.

This should address georchestra issue #1265

Tests: manual, but should not harm the CI.