mapfishapp: re-implement GeoJSON Feature encoding in a streaming way
The encoding of FeatureCollections to GeoJSON was
partially streaming, by means of using a FeatureCollectionEncoder
as the "features" property on a Map<String, Object>. This
encoder, being a JSONStreamAware, encodes feature by feature
to the output stream, but it did so in a not so streaming
fashion: by first converting each Feature to a String and then
writing the String verbatim to the output stream.

Due to the way most of the GeoTools GeoJSON writer is currently
implemented, that resulted in an explosion of objects on the
heap for each Feature (java.io.Writer, String, and other object
instances).

This patch makes FeatureEncoder also a org.json.simple.JSONStreamAware
instead of a simple org.json.simple.JSONAware, which merely converts
a Feature to String, to instead write to a org.json.JSONWriter
directly without incurring in the perf and memory overhead.

To make it optimal, the only missing piece would be adding
the ability to also stream out Geometry serialization, which
is currently handled by the GeoTools GeometryJSON class and builds
a java.util.Map with nested properties, Lists, and Maps to then
serialize it as String.